requirements.txt
numpy
pandas
Pillow

pickle
plotly
networkx
streamlit

tqdm
scipy
matplotlib

